Fixed lock LEDs problem
